+++ toki/toki/core/utils/string_dumpers.h
@@ -13,7 +13,7 @@ struct StringDumper {};
 
 template <>
 struct StringDumper<char*> {
-	static constexpr u32 dump_to_string(char* out, const char* arg) {
+	static constexpr u64 dump_to_string(char* out, const char* arg) {
 		u64 length = toki::strlen(arg);
 		toki::memcpy(out, arg, toki::strlen(arg));
 		return length;
@@ -24,7 +24,7 @@ static_assert(CHasDumpToString<char*>);
 
 template <>
 struct StringDumper<char> {
-	static constexpr u32 dump_to_string(char* out, const char arg) {
+	static constexpr u64 dump_to_string(char* out, const char arg) {
 		out[0] = arg;
 		return 1;
 	}
@@ -34,7 +34,7 @@ static_assert(CHasDumpToString<char>);
 
 template <>
 struct StringDumper<bool> {
-	static constexpr u32 dump_to_string(char* out, const bool arg) {
+	static constexpr u64 dump_to_string(char* out, const bool arg) {
 		if (arg) {
 			toki::memcpy(out, TRUE_STR, strlen(TRUE_STR));
 			return strlen(TRUE_STR);
@@ -50,7 +50,7 @@ static_assert(CHasDumpToString<bool>);
 template <typename T>
 	requires CIsIntegral<T>
 struct StringDumper<T> {
-	static constexpr u32 dump_to_string(char* out, const T arg) {
+	static constexpr u64 dump_to_string(char* out, const T arg) {
 		return itoa(out, remove_ref(arg));
 	}
 };
